Tower collision detection and shooting
Towers can now target enemies using a quad tree. Tower then picks the
first enemy from the enemies in range and begins shooting it. Added
lives counter and lives go down when enemies reach the end of the map.

+++ Source/Framework/position.cpp
@@ -5,7 +5,7 @@ Position::Position()
 {
 }
 
-Position::Position(int xPos, int yPos)
+Position::Position(float xPos, float yPos)
 {
 	m_x = xPos;
 	m_y = yPos;
